requirements.txt
Flask
Flask-Cors
python-dotenv
openai>=1.0.0
